Fix: renamed Protected to ProtectedRoute / Delete: react navigate from appjs
